vMU/src/requirements.txt
numpy
psutil
pyasn1
PyYAML
setuptools
scipy